# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "operationId": "TasksController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new task",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TasksController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "incidentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedToId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all tasks with optional filters",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/my-tasks": {
      "get": {
        "operationId": "TasksController_getMyTasks",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns tasks assigned to current user"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get tasks assigned to current user",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/statistics": {
      "get": {
        "operationId": "TasksController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get task statistics",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "get": {
        "operationId": "TasksController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a task by ID",
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "TasksController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a task",
        "tags": [
          "tasks"
        ]
      },
      "delete": {
        "operationId": "TasksController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a task (Admin only)",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}/accept": {
      "post": {
        "operationId": "TasksController_acceptTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task accepted and status changed to IN_PROGRESS"
          },
          "400": {
            "description": "Task is not in PENDING status"
          },
          "403": {
            "description": "Not assigned to this task"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept a task (Field Officer)",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}/survey": {
      "post": {
        "operationId": "TasksController_submitSurveyData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSurveyDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Survey data submitted and status changed to SURVEYED"
          },
          "400": {
            "description": "Task is not in IN_PROGRESS status"
          },
          "403": {
            "description": "Not assigned to this task"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit field survey data",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/surveys/{id}": {
      "get": {
        "operationId": "SurveyController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The found survey",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyDto"
                }
              }
            }
          },
          "404": {
            "description": "Survey not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a survey by ID",
        "tags": [
          "Surveys"
        ]
      }
    },
...
    "/field-officer/surveys": {
      "post": {
        "description": "Field officers use this endpoint to submit survey data collected in the field, including GPS coordinates and photos",
        "operationId": "FieldOfficerSurveyController_submitSurvey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FieldOfficerSurveyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Survey submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOfficerSurveyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Field Officer role required"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit initial survey from field",
        "tags": [
          "Field Officer - Surveys"
        ]
      }
    },
    "/field-officer/surveys/my-surveys": {
      "get": {
        "description": "Retrieve all surveys submitted by the current field officer",
        "operationId": "FieldOfficerSurveyController_getMySurveys",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of submitted surveys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FieldOfficerSurveyResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my submitted surveys",
        "tags": [
          "Field Officer - Surveys"
        ]
      }
    },
    "/field-officer/surveys/{id}": {
      "get": {
        "description": "Retrieve a specific survey by its ID",
        "operationId": "FieldOfficerSurveyController_getSurvey",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Survey details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FieldOfficerSurveyResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Survey not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get survey by ID",
        "tags": [
          "Field Officer - Surveys"
        ]
      }
    },
...
  },
  "info": {
    "title": "Guardian Route API",
    "description": "\n      Guardian Route - ระบบบริหารจัดการภัยพิบัติ\n      ตำบลเวียง อำเภอฝาง จังหวัดเชียงใหม่\n      \n      ## Authentication\n      Most endpoints require JWT Bearer token authentication.\n      Use the /api/auth/login endpoint to obtain a token.\n      \n      ## Roles\n      - DEVELOPER: Full system access\n      - ADMIN: System administration\n      - SUPERVISOR: Command and control\n      - EXECUTIVE: Strategic oversight\n      - FIELD_OFFICER: Field operations\n      \n      ## Base URL\n      - Development: http://localhost:3001/api\n      - Production: https://api.guardian-route.example.com/api\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Guardian Route Team",
      "url": "https://github.com/jetci/Guardian-Route",
      "email": "support@guardian-route.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.guardian-route.example.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Task title"
          },
          "description": {
            "type": "string",
            "description": "Task description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "description": "Task priority"
          },
          "incidentId": {
            "type": "string",
            "description": "Related incident ID"
          },
          "assignedToId": {
            "type": "string",
            "description": "Assigned user ID"
          },
          "dueDate": {
            "type": "string",
            "description": "Due date"
          }
        },
        "required": [
          "title",
          "priority",
          "incidentId"
        ]
      },
      "GeoJSONPoint": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Point",
            "description": "GeoJSON type"
          },
          "coordinates": {
            "example": [
              98.9817,
              19.9167
            ],
            "description": "[longitude, latitude]",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "coordinates"
        ]
      },
      "GeoJSONPolygon": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Polygon",
            "description": "GeoJSON type"
          },
          "coordinates": {
            "example": [
              [
                [
                  98.9817,
                  19.9167
                ],
                [
                  98.9827,
                  19.9167
                ],
                [
                  98.9827,
                  19.9177
                ],
                [
                  98.9817,
                  19.9177
                ],
                [
                  98.9817,
                  19.9167
                ]
              ]
            ],
            "description": "Array of linear rings (array of coordinates) following GeoJSON spec",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "coordinates"
        ]
      },
      "UpdateSurveyDataDto": {
        "type": "object",
        "properties": {
          "surveyLocation": {
            "description": "Survey location (GPS point)",
            "example": {
              "type": "Point",
              "coordinates": [
                98.9817,
                19.9167
              ]
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPoint"
              }
            ]
          },
          "surveyArea": {
            "description": "Survey area (drawn polygon)",
            "example": {
              "type": "Polygon",
              "coordinates": [
                [
                  [
                    98.9817,
                    19.9167
                  ],
                  [
                    98.9827,
                    19.9167
                  ],
                  [
                    98.9827,
                    19.9177
                  ],
                  [
                    98.9817,
                    19.9177
                  ],
                  [
                    98.9817,
                    19.9167
                  ]
                ]
              ]
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPolygon"
              }
            ]
          },
          "surveyNotes": {
            "type": "string",
            "description": "Survey notes and observations",
            "example": "Water level increased by 15cm since last visit. Noted minor landslide near village entrance."
          },
          "surveyPhotos": {
            "description": "Array of photo URLs",
            "example": [
              "https://cdn.guardian-route.example.com/uploads/surveys/IMG_001.jpg",
              "https://cdn.guardian-route.example.com/uploads/surveys/IMG_002.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Task title"
          },
          "description": {
            "type": "string",
            "description": "Task description"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ],
            "description": "Task priority"
          },
          "incidentId": {
            "type": "string",
            "description": "Related incident ID"
          },
          "assignedToId": {
            "type": "string",
            "description": "Assigned user ID"
          },
          "dueDate": {
            "type": "string",
            "description": "Due date"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "SURVEYED",
              "COMPLETED",
              "CANCELLED"
            ],
            "description": "Task status"
          },
          "surveyLocation": {
            "type": "object",
            "description": "Survey location (GeoJSON Point)"
          },
          "surveyArea": {
            "type": "object",
            "description": "Survey area (GeoJSON Polygon)"
          },
          "surveyNotes": {
            "type": "string",
            "description": "Survey notes"
          },
          "surveyPhotos": {
            "description": "Survey photos URLs",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SurveyTemplateDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "object"
          },
          "formSchema": {
            "type": "object",
            "description": "JSON schema of the survey form",
            "additionalProperties": true
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "formSchema",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "SurveyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the survey"
          },
          "incidentId": {
            "type": "object",
            "description": "ID of the Incident this survey is related to"
          },
          "villageId": {
            "type": "object",
            "description": "ID of the Village this survey is related to"
          },
          "templateId": {
            "type": "string",
            "description": "ID of the Survey Template used"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ],
            "description": "Current status of the survey"
          },
          "completedAt": {
            "type": "object",
            "description": "Date and time when the survey was completed"
          },
          "polygon": {
            "type": "object",
            "description": "GeoJSON Polygon for the survey area (optional)"
          },
          "createdById": {
            "type": "string",
            "description": "ID of the user who created the survey"
          },
          "template": {
            "description": "Survey Template details",
            "allOf": [
              {
                "$ref": "#/components/schemas/SurveyTemplateDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "templateId",
          "status",
          "createdById"
        ]
      },
...
      "GPSCoordinatesDto": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "example": 19.9167,
            "description": "Latitude"
          },
          "lng": {
            "type": "number",
            "example": 99.2333,
            "description": "Longitude"
          },
          "accuracy": {
            "type": "number",
            "example": 10.5,
            "description": "GPS accuracy in meters"
          },
          "altitude": {
            "type": "number",
            "example": 350.5,
            "description": "Altitude in meters"
          }
        },
        "required": [
          "lat",
          "lng"
        ]
      },
      "FieldOfficerSurveyDto": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string",
            "example": "uuid-of-task",
            "description": "Related task ID"
          },
          "incidentId": {
            "type": "string",
            "example": "uuid-of-incident",
            "description": "Related incident ID"
          },
          "villageId": {
            "type": "string",
            "example": "uuid-of-village",
            "description": "Village ID (optional if village name is provided)"
          },
          "villageName": {
            "type": "string",
            "example": "บ้านเต๋าดิน หมู่ 3",
            "description": "Village name"
          },
          "disasterType": {
            "type": "string",
            "example": "น้ำท่วม",
            "description": "Disaster type (น้ำท่วม, ดินถ่ม, ไฟป่า, etc.)"
          },
          "severity": {
            "type": "number",
            "example": 3,
            "description": "Severity level (1-5)",
            "minimum": 1,
            "maximum": 5
          },
          "estimatedHouseholds": {
            "type": "number",
            "example": 25,
            "description": "Estimated affected households"
          },
          "notes": {
            "type": "string",
            "example": "น้ำท่วมขังบริเวณถนนสายหลัก ระดับน้ำประมาณ 50 ซม. ประชาชนต้องการความช่วยเหลือด้านอาหารและน้ำดื่ม",
            "description": "Survey notes and observations"
          },
          "gpsLocation": {
            "description": "GPS coordinates of survey location",
            "allOf": [
              {
                "$ref": "#/components/schemas/GPSCoordinatesDto"
              }
            ]
          },
          "photoUrls": {
            "example": [
              "https://example.com/photo1.jpg",
              "https://example.com/photo2.jpg"
            ],
            "description": "Array of photo URLs (uploaded separately)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "polygon": {
            "type": "object",
            "example": {
              "type": "Polygon",
              "coordinates": [
                [
                  [
                    99.8833,
                    19.9167
                  ],
                  [
                    99.885,
                    19.9167
                  ],
                  [
                    99.885,
                    19.915
                  ],
                  [
                    99.8833,
                    19.915
                  ],
                  [
                    99.8833,
                    19.9167
                  ]
                ]
              ]
            },
            "description": "GeoJSON polygon of affected area (drawn on map)"
          },
          "areaSize": {
            "type": "number",
            "example": 2.5,
            "description": "Area size in km² (calculated from polygon)"
          },
          "additionalData": {
            "type": "object",
            "example": {
              "weather": "ฝนตก",
              "temperature": "28°C",
              "visibility": "ดี"
            },
            "description": "Additional survey data (flexible JSON)"
          }
        },
        "required": [
          "villageName",
          "disasterType",
          "severity",
          "estimatedHouseholds",
          "notes",
          "gpsLocation"
        ]
      },
      "FieldOfficerSurveyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-of-survey"
          },
          "fieldOfficerId": {
            "type": "string",
            "example": "uuid-of-user"
          },
          "taskId": {
            "type": "string",
            "example": "uuid-of-task"
          },
          "incidentId": {
            "type": "string",
            "example": "uuid-of-incident"
          },
          "villageId": {
            "type": "string",
            "example": "uuid-of-village"
          },
          "villageName": {
            "type": "string",
            "example": "บ้านเต๋าดิน หมู่ 3"
          },
          "disasterType": {
            "type": "string",
            "example": "น้ำท่วม"
          },
          "severity": {
            "type": "number",
            "example": 3
          },
          "estimatedHouseholds": {
            "type": "number",
            "example": 25
          },
          "notes": {
            "type": "string",
            "example": "Survey notes..."
          },
          "gpsLocation": {
            "$ref": "#/components/schemas/GPSCoordinatesDto"
          },
          "photoUrls": {
            "example": [
              "url1",
              "url2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "polygon": {
            "type": "object",
            "example": {
              "type": "Polygon",
              "coordinates": [
                [
                  [
                    99.8833,
                    19.9167
                  ]
                ]
              ]
            }
          },
          "areaSize": {
            "type": "number",
            "example": 2.5
          },
          "additionalData": {
            "type": "object"
          },
          "submittedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-11-29T14:30:00Z"
          },
          "status": {
            "type": "string",
            "example": "SUBMITTED"
          }
        },
        "required": [
          "id",
          "fieldOfficerId",
          "taskId",
          "incidentId",
          "villageId",
          "villageName",
          "disasterType",
          "severity",
          "estimatedHouseholds",
          "notes",
          "gpsLocation",
          "photoUrls",
          "additionalData",
          "submittedAt",
          "status"
        ]
      },
...
    }
  }
}